Initial commit
The files in the ChessEngine and GUI folders presumably work well, but everything in the Bot folder still needs work. To play a game, run Launcher.py in the GUI folder. While the engine will recognize if a game is a draw by threefold repetition or by the 50-move rule, the GUI itself will not prevent the player from making moves. Currently the B key will perform random moves for as long as it is held down, which can easily be changed to only activating on key press by uncommenting some code. In the future, the idea is to have this use the bot to make a move. The U key will undo moves as long as it is held, which also be be made to only activate on key press by uncommenting code.